Add custom types & remove any when possible
