+++ frontend/{{cookiecutter.project_slug}}/src/views/HomeView.vue
@@ -239,7 +239,7 @@ import SelectField from '@/components/common/form/SelectField.vue'
 import CurrencyField from '@/components/common/form/CurrencyField.vue'
 import DatePicker from '@/components/common/form/DatePicker.vue'
 import formValidation from '@/helpers/form-validation.ts'
-import { SnackBar } from '@/api/config.ts'
+import { SnackBar } from '@/types/api-types.ts'
 import { ref } from 'vue'
 import { useApis } from '@/composables/use-apis.ts'
 import { wrapper } from '@/composables/use-api-wrapper.ts'
